Text Viewer: Use int instead of size_t when possible.

+++ src/3rdparty/text_viewer/text_viewer_demo.cpp
@@ -51,7 +51,7 @@ int main() {
     return 0;
 })");
 
-            viewer.set_text(tv::string_view(text.data(), text.size()));
+            viewer.set_text(tv::string_view(text.data(), (int)text.size()));
         }
 
         {
